+++ packages/eslint-plugin-fiori-tools/jest.config.js
@@ -1,7 +1,7 @@
 const config = require('../../jest.base');
 module.exports = {
     ...config,
-    // Coverage is handled by c8 wrapper for worker thread support
-    collectCoverage: false,
+    // coverageProvider: 'v8' overrides jest.base default ('babel'); collectCoverage: true is inherited from jest.base
+    coverageProvider: 'v8',
     setupFiles: ['<rootDir>/test/global-setup.ts']
 }