+++ packages/fiori-app-sub-generator/test/int/fiori-elements/headless.test.ts
@@ -175,6 +175,72 @@ describe('Headless generation', () => {
         cleanTestDir(join(testDir, testProjectNameNoVers));
     });
 
+    it('LROP v4 with externalServices - external service files are written', async () => {
+        testProjectName = 'lrop_v4_with_vh';
+        expectedOutputPath = join(
+            __dirname,
+            EXPECTED_OUTPUT_DIR_NAME,
+            'headless',
+            testProjectName,
+            projectAccess.DirName.Webapp,
+            projectAccess.DirName.LocalService
+        );
+        const fixturesHeadlessDir = join(__dirname, './fixtures/headless');
+
+        const appConfigJson = JSON.stringify({
+            version: '0.2',
+            floorplan: 'FE_LROP',
+            project: {
+                title: 'Project\'s "Title"',
+                description: 'Test \'Project\' "Description"',
+                namespace: 'testNameSpace',
+                ui5Version: '1.84.0',
+                localUI5Version: '1.82.2',
+                targetFolder: testDir,
+                name: testProjectName,
+                sapux: true,
+                enableEslint: false
+            },
+            service: {
+                servicePath: '/sap/opu/odata4/dmo/sb_travel_mduu_o4/srvd/dmo/sd_travel_mduu/0001/',
+                host: 'https://sap-ux-mock-services-v4-feop.cfapps.us10.hana.ondemand.com',
+                externalServices: [
+                    {
+                        type: 'value-list',
+                        target: 'Travel/AgencyID',
+                        metadata: join(fixturesHeadlessDir, 'agency_vh_metadata.xml'),
+                        path: '/sap/opu/odata4/sap/agency_vh_service/srvd/sap/agency_vh_service/0001/',
+                        entityData: join(fixturesHeadlessDir, 'agency_entity_data.json')
+                    },
+                    {
+                        type: 'code-list',
+                        collectionPath: 'Currencies',
+                        metadata:
+                            '<?xml version="1.0" encoding="utf-8"?><edmx:Edmx Version="4.0" xmlns:edmx="http://docs.oasis-open.org/odata/ns/edmx"><edmx:DataServices><Schema Namespace="Currencies" xmlns="http://docs.oasis-open.org/odata/ns/edm"></Schema></edmx:DataServices></edmx:Edmx>',
+                        path: '/sap/opu/odata4/sap/currencies/srvd/sap/currencies/0001/'
+                    },
+                    {
+                        type: 'code-list',
+                        collectionPath: 'Units',
+                        metadata: join(fixturesHeadlessDir, 'common_metadata.xml'),
+                        path: '/sap/opu/odata4/sap/common/srvd/sap/common/0001/'
+                    }
+                ]
+            },
+            entityConfig: {
+                mainEntity: { entityName: 'Travel' },
+                navigationEntity: { EntitySet: 'Booking', Name: '_Booking' }
+            }
+        });
+
+        await runHeadlessGen(appConfigJson, 'travel_v4');
+        expect(
+            join(testDir, testProjectName, projectAccess.DirName.Webapp, projectAccess.DirName.LocalService)
+        ).toMatchFolder(expectedOutputPath);
+
+        cleanTestDir(join(testDir, testProjectName));
+    });
+
     it('LROP v4 CAP', async () => {
         testProjectName = 'lrop_v4_cap';
         expectedOutputPath = join(__dirname, EXPECTED_OUTPUT_DIR_NAME, testProjectName);