+++ packages/annotation-generator/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sap-ux/annotation-generator
 
+## 0.4.51
+
+### Patch Changes
+
+- Updated dependencies [52f6549]
+    - @sap-ux/fiori-annotation-api@0.9.44
+
+## 0.4.50
+
+### Patch Changes
+
+- Updated dependencies [03d3ea1]
+    - @sap-ux/project-access@1.35.21
+    - @sap-ux/fiori-annotation-api@0.9.43
+
 ## 0.4.49
 
 ### Patch Changes